requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-Mail==0.9.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Assets==2.1.0
email-validator==2.1.0
python-dotenv==1.0.1
Werkzeug==2.3.7
SQLAlchemy==1.4.50
wtforms==3.1.1
jsmin==3.0.1
cssmin==0.2.0
psycopg2-binary==2.9.10 